refactor: Rewrite code, separate into modules. Add abstract class
