+++ src/bot/__init__.py
@@ -1,0 +1,3 @@
+from .bot_handler import RecipeBot
+
+__all__ = ['RecipeBot']
